Moved TypeDescriptionUtility to the Common folder/namespace. Updated the TypeDescriptionUtilitySpecs.cs class correspondingly. All actual uses of TypeDescriptionUtility already import this namespace so no other changes are needed.
