+++ Tests/FluentAssertions.Specs/Common/TypeDescriptionUtilitySpecs.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
+using FluentAssertions.Common;
+
 using Xunit;
 
 namespace FluentAssertions.Specs
